♻️ datasource·repository 의 타입 선언을 entity 로 이관
- chat-datasource 의 Request/Params 타입 5개(CreateChatRoomRequest,
  GetChatRoomParams, GetChatRoomsParams, GetMessagesParams,
  SendMessageRequest) → chat.entity 로 이동
- movie-repository 의 CGVTheater·CGVTheaterList 인터페이스 → movie.entity
  로 이동
- chat/index.ts 도 entity 재수출에 맞춰 정리

타입은 entity 단일 출처 원칙. datasource·repository 는 로직만 담는다.
