+++ src/modules/chat/chat-datasource.ts
@@ -4,37 +4,12 @@ import type {
   ChatRoomEntity,
   ChatRoomsResponseEntity,
   MessagesResponseEntity,
+  CreateChatRoomRequest,
+  GetChatRoomParams,
+  GetChatRoomsParams,
+  GetMessagesParams,
+  SendMessageRequest,
 } from './chat.entity'
-
-export interface CreateChatRoomRequest {
-  memberIds: number[]
-  roomName: string
-  type: 'direct' | 'group'
-}
-
-export interface GetChatRoomParams {
-  chatRoomId: string
-  userId: number
-}
-
-export interface GetChatRoomsParams {
-  userId: number
-  page?: number
-  pageSize?: number
-}
-
-export interface GetMessagesParams {
-  chatRoomId: string
-  userId: number
-  page?: number
-  pageSize?: number
-}
-
-export interface SendMessageRequest {
-  chatRoomId: string
-  senderId: number
-  content: string
-}
 
 export class ChatDatasource {
   private token?: string
