+++ src/modules/chat/index.ts
@@ -1,10 +1,3 @@
 export * from './chat.entity'
 export { ChatDatasource } from './chat-datasource'
-export type {
-  CreateChatRoomRequest,
-  GetChatRoomParams,
-  GetChatRoomsParams,
-  GetMessagesParams,
-  SendMessageRequest,
-} from './chat-datasource'
 export { ChatRepository } from './chat-repository'
