+++ src/modules/movie/movie-repository.ts
@@ -1,6 +1,7 @@
 import { MovieDatasource } from './movie-datasource'
 import {
   AverageMovieScore,
+  CGVTheaterList,
   Movie,
   Score,
   assertAverageMovieScore,
@@ -8,21 +9,6 @@ import {
   assertScore,
 } from './movie.entity'
 
-export interface CGVTheaterList {
-  theaters: CGVTheater[]
-}
-export interface CGVTheater {
-  id: number
-  name: string
-  region: string
-  address: string
-  phone: string
-  website: string
-  latitude: number
-  longitude: number
-  createdAt: string
-  updatedAt: string
-}
 export class MovieRepository {
   private datasource: MovieDatasource
   constructor(
